Refactor: pull savegame to vegadisk
