+++ libraries/cmd/unit_functions_generic.cpp
@@ -26,7 +26,7 @@
 #include "gfx_generic/cockpit_generic.h"
 #include "root_generic/faction_generic.h"
 #include "cmd/ai/communication.h"
-#include "root_generic/savegame.h"
+#include "vegadisk/savegame.h"
 #include "root_generic/xml_support.h"
 #include "cmd/unit_util.h"
 #include "src/universe_util.h"
