+++ libraries/cmd/unit_util_generic.cpp
@@ -31,7 +31,7 @@
 #include "gfx_generic/cockpit_generic.h"
 #include "src/universe_util.h"
 #include "cmd/ai/docking.h"
-#include "root_generic/savegame.h"
+#include "vegadisk/savegame.h"
 #include "cmd/planet.h"
 #include "root_generic/faction_generic.h"
 #include "cmd/ai/fire.h"
