+++ libraries/root_generic/load_mission.cpp
@@ -42,7 +42,7 @@
 #include "cmd/ai/missionscript.h"
 #include "cmd/script/flightgroup.h"
 #include "src/python/python_class.h"
-#include "root_generic/savegame.h"
+#include "vegadisk/savegame.h"
 #include "src/save_util.h"
 #include "root_generic/load_mission.h"
 #include "src/universe.h"
