+++ libraries/vegadisk/savegame.cpp
@@ -38,7 +38,7 @@
 #include <vector>
 #include <string>
 #include "root_generic/vs_globals.h"
-#include "root_generic/savegame.h"
+#include "vegadisk/savegame.h"
 #include "root_generic/load_mission.h"
 #include <algorithm>
 #include "cmd/script/mission.h"
