+++ libraries/vegadisk/savegame.h
